Moved AWS mocks into a custom package.
This mocks are only useful in some kind of tests dealing with AWS calls - import this new package in your tests if you need these AWS resources mocks.
